refactor: move ZoomTransmissionFraction workflows to esssans
